music-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-arango==7.6.1
pydantic==2.9.2
python-dotenv==1.0.1
